pyproject.toml
[project]
name = "backend-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-google-genai>=2.1.9",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
